[License Cleanup]: Region.Framework.Scenes
- Cleanup license text
- Cleanup comments
- Remove commented out code to ensure binary files do not get bloated
- Add the { } in the proper statements to make sure build systems make a clean exit from statements.
- Resharp files

NOTE 1:

By cleaning out the commented out code this helps to reduce the chances of binary files getting bloated.  It will also help make sure that the real problems found by Coverity and Codacy scans are actually issues and not commented out code.  The Commented out code does not need to be in the file as each commit to the repository shows the changes made in that commit and it is stored in the commit history.